requirements.txt
requests
arxiv
tqdm
watchdog
pdftotext
boto
lxml
mmh3
pdfparser.six
ujson
